backend/requirements/base.txt
django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-split-settings==1.2.0
django-environ==0.11.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0